fix(editor): use distance scale factor to calculated default travel time
fixes #197
