+++ lib/editor/components/EditorFeedSourcePanel.js
@@ -22,8 +22,6 @@ export default class EditorFeedSourcePanel extends Component {
 
   messages = getComponentMessages('EditorFeedSourcePanel')
 
-  state = { expanded: false }
-
   componentWillMount () {
     this.props.getSnapshots(this.props.feedSource)
   }
