+++ lib/editor/components/pattern/CalculateDefaultTimesForm.js
@@ -3,7 +3,7 @@ import React, {Component, PropTypes} from 'react'
 import { Button, FormGroup, InputGroup, Form, FormControl, ControlLabel } from 'react-bootstrap'
 
 import MinuteSecondInput from '../MinuteSecondInput'
-import {straightLineDistancesBetweenStopAnchors} from '../../util/map'
+import {getDistanceScaleFactor, straightLineDistancesBetweenStopAnchors} from '../../util/map'
 
 const DEFAULT_SPEED = 20 // km/hr
 
@@ -25,6 +25,10 @@ export default class CalculateDefaultTimesForm extends Component {
     const patternStops = [...activePattern.patternStops]
     const convertedSpeed = speed * 1000 / 60 / 60 // km/hr -> m/s
     const straightLineDistances = straightLineDistancesBetweenStopAnchors(controlPoints)
+    // Get the distance scale factor (pattern length / total shape dist traveled).
+    const distScaleFactor = activePattern.shape
+      ? getDistanceScaleFactor(activePattern.shapePoints)
+      : 1
     for (var i = 0; i < patternStops.length; i++) {
       let interStopDistance = 0
       patternStops[i].defaultDwellTime = dwellTime
@@ -50,7 +54,8 @@ export default class CalculateDefaultTimesForm extends Component {
             negative or zero (${interStopDistance} meters). Defaulting travel time to zero.`,
             patternStops[i - 1], patternStops[i])
         } else {
-          patternStops[i].defaultTravelTime = Math.floor(interStopDistance / convertedSpeed)
+          const interStopDistanceInMeters = interStopDistance * distScaleFactor
+          patternStops[i].defaultTravelTime = Math.floor(interStopDistanceInMeters / convertedSpeed)
         }
       }
     }
