+++ lib/editor/components/timetable/TimetableGrid.js
@@ -281,8 +281,7 @@ export default class TimetableGrid extends Component {
         placeholder={col.placeholder}
         renderTime={isTimeFormat(col.type)}
         rowIndex={rowIndex}
-        style={style}
-        />
+        style={style} />
     )
   }
 
@@ -350,8 +349,8 @@ export default class TimetableGrid extends Component {
     return col.type === 'ARRIVAL_TIME' && hideDepartureTimes
       ? col.width * 2
       : col
-      ? col.width
-      : 90
+        ? col.width
+        : 90
   }
 
   /**
