+++ lib/editor/selectors/index.js
@@ -17,6 +17,7 @@ import {getTableById, getEntityName} from '../util/gtfs'
 import {POINT_TYPE} from '../constants'
 import {
   generateControlPointsFromPatternStops,
+  getDistanceScaleFactor,
   getLineSlices,
   newControlPoint,
   projectStopOntoLine
@@ -273,31 +274,6 @@ export const getControlPoints = createSelector(
     }
   }
 )
-
-/**
- * Get a scale factor to multiply shape dist traveled values by to get distance
- * traveled in meters. The shapeDistTraveled field is intended to be unitless
- * (see GTFS spec: https://developers.google.com/transit/gtfs/reference/#shapestxt).
- * But the units *should* be the same between stop_times.txt, which is used to derive
- * the pattern_stops values, and shapes.txt So, we use the last shapeDistTraveled
- * value for the pattern and the pattern length in meters to come up with a
- * scale factor to multiply by all of the pattern stop shapeDistTraveled values.
- */
-function getDistanceScaleFactor (shapePoints) {
-  const patternLine = lineString(coordinatesFromShapePoints(shapePoints))
-  const patternLengthInMeters = lineDistance(patternLine, 'meters')
-  const totalShapePointsDistTraveled = shapePoints[shapePoints.length - 1].shapeDistTraveled
-  // If the shape points have no shape dist traveled values, default the scale
-  // factor to one.
-  let distScaleFactor
-  if (totalShapePointsDistTraveled) {
-    distScaleFactor = patternLengthInMeters / totalShapePointsDistTraveled
-  } else {
-    console.warn(`No shape dist traveled value found for final shape point. Distance scale factor defaulting to 1.`)
-    distScaleFactor = 1
-  }
-  return distScaleFactor
-}
 
 /**
  * Gets the control points from the list of shape points. Any shape points that
