+++ lib/editor/util/map.js
@@ -15,6 +15,7 @@ import {POINT_TYPE} from '../constants'
 import {getSegment} from '../../scenario-editor/utils/valhalla'
 import {generateUID} from '../../common/util/util'
 import {getConfigProperty} from '../../common/util/config'
+import {coordinatesFromShapePoints} from './objects'
 import {reverseEsri as reverse} from '../../scenario-editor/utils/reverse'
 
 import type {
@@ -28,7 +29,8 @@ import type {
   LatLng,
   Pattern,
   PatternStop,
-  ShapePoint
+  ShapePoint,
+  StopTime
 } from '../../types'
 
 type MapLayer = {
@@ -612,7 +614,7 @@ export function resequenceShapePoints (shapePoint: ShapePoint, index: number): S
 /**
  * Map function that resets stop sequence values for pattern stops.
  */
-export function resequenceStops (stop: PatternStop, index: number): PatternStop {
+export function resequenceStops (stop: PatternStop | StopTime, index: number): PatternStop {
   return {
     ...stop,
     stopSequence: index
@@ -893,4 +895,29 @@ export function projectStopOntoLine (stop: GtfsStop, line: any): any {
   // Distance returned from nearestPointOnLine defaults to km.
   const distanceInMeters = insertPoint.properties.location * 1000
   return {insertPoint, insertIndex, distanceInMeters}
+}
+
+/**
+ * Get a scale factor to multiply shape dist traveled values by to get distance
+ * traveled in meters. The shapeDistTraveled field is intended to be unitless
+ * (see GTFS spec: https://developers.google.com/transit/gtfs/reference/#shapestxt).
+ * But the units *should* be the same between stop_times.txt, which is used to derive
+ * the pattern_stops values, and shapes.txt So, we use the last shapeDistTraveled
+ * value for the pattern and the pattern length in meters to come up with a
+ * scale factor to multiply by all of the pattern stop shapeDistTraveled values.
+ */
+export function getDistanceScaleFactor (shapePoints: Array<ShapePoint>): number {
+  const patternLine = lineString(coordinatesFromShapePoints(shapePoints))
+  const patternLengthInMeters = lineDistance(patternLine, 'meters')
+  const totalShapePointsDistTraveled = shapePoints[shapePoints.length - 1].shapeDistTraveled
+  // If the shape points have no shape dist traveled values, default the scale
+  // factor to one.
+  let distScaleFactor
+  if (totalShapePointsDistTraveled) {
+    distScaleFactor = patternLengthInMeters / totalShapePointsDistTraveled
+  } else {
+    console.warn(`No shape dist traveled value found for final shape point. Distance scale factor defaulting to 1.`)
+    distScaleFactor = 1
+  }
+  return distScaleFactor
 }
